table-cards-view (cljs) instead of table-view (js)
since we're not using canvas for the time being, i don't see why the
table-(view-)cards component has to be in vanilla js. i actually started
running into issue with the vanilla js component wanting to access cljx
methods and not being sure how to do it. (now that i'm writing this,
maybe it wouldn't be so hard. i could use `js->clj` as outlined in
[1]). anyway, i just feel like we should stay in cljs if possible. it
will reduce the amount of plumbing.

[1] http://www.spacjer.com/blog/2014/09/12/clojurescript-javascript-interop/