split up input and output channels
chord's implementation of channels might (?) be a bit
disingenuous(?). it seems to know that the server should always send to
the client and the client should always send to the server. even though
you're sending to and receiving from the same channel, i expect there
are actually two channels underneath(?).

the problem i was running into with plugging in an AI is that i wanted
to still use a core.async channel, but not an actual websocket. i
realized both ends were listening on and sending to the same socket, so
the AI was receiving its own messages(?).

all conjecture at this point. but it seems like i'm going to need one
channel to send from the AI to the server, and another to send back from
the server to the AI. i think chord is hiding this implementation detail
from me so it sets the same channel as both the input and the output.